requirements.txt
azure-ai-formrecognizer==3.3.2
azure-cognitiveservices-vision-computervision==0.9.0
python-dotenv==1.0.0
pandas==2.0.3
python-multipart==0.0.6
fastapi==0.95.2
uvicorn==0.22.0
